subscriber/requirements.txt
paho-mqtt==2.1.0
requests==2.32.3
jsonschema==4.23.0
python-dotenv==1.1.0